[FEAT] #70 - test code 작성을 위해 repository 를 interface 로 전환
